backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.10.0
pydantic-settings>=2.6.0
google-cloud-storage>=2.14.0
google-cloud-dlp>=3.15.0
google-cloud-documentai>=2.24.0
google-cloud-aiplatform==1.68.0
google-cloud-secret-manager>=2.18.0
google-cloud-logging==3.10.0
cloud-sql-python-connector[pg8000]>=1.7.0
pg8000>=1.30.5
sqlalchemy>=2.0.23
PyMuPDF>=1.23.21
pikepdf>=8.11.2
Pillow>=10.2.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
